Finish ability to message for opportunities to collaborate
Fixes: https://github.com/Hylozoic/hylo-evo/issues/1179
