+++ src/components/Widget/OpportunitiesToCollaborateWidget/OpportunitiesToCollaborateWidget.js
@@ -1,8 +1,10 @@
 import React from 'react'
+import { useDispatch } from 'react-redux'
 import Icon from 'components/Icon'
 import useRouter from 'hooks/useRouter'
 import useEnsureCurrentGroup from 'hooks/useEnsureCurrentGroup'
 import { getFarmOpportunities } from 'store/selectors/farmExtensionSelectors'
+import { messageGroupModerators } from '../Widget.store'
 
 import './OpportunitiesToCollaborate.scss'
 
@@ -12,15 +14,12 @@ export default function OpportunitiesToCollaborateWidget () {
   const opportunities = getFarmOpportunities(group)
   const moderatorIds = group.moderators && group.moderators.map((mod) => mod.id)
   const hasMods = moderatorIds.length !== 0
+  const dispatch = useDispatch()
 
   return (
     <div styleName='opportunities-to-collaborate-container'>
       {opportunities && opportunities.length > 0 && opportunities.map((opportunity) => {
-        const query = new URLSearchParams('')
-        if (hasMods) {
-          query.append('participants', moderatorIds.join(','))
-          query.append('prompt', opportunity)
-        }
+        const prompt = `Hi there ${group.name}, I'd like to talk about ${promptLookup[opportunity]}.`
         return (
           <div styleName='collab-item' key={opportunity}>
             <Icon styleName='collab-icon' blue name={opportunity.charAt(0).toUpperCase() + opportunity.slice(1)} />
@@ -28,7 +27,12 @@ export default function OpportunitiesToCollaborateWidget () {
               <div styleName='collab-title'>{collabTitle[opportunity]}</div>
               <div styleName='collab-text'>{collabText(group)[opportunity]}</div>
             </div>
-            {hasMods && <Icon name='Messages' blue={hasMods} styleName={`collab-icon${hasMods ? ' cursor-pointer' : ''}`} onClick={hasMods ? () => push(`/messages/new${hasMods ? '?' + query.toString() : ''}`) : null} />}
+            <Icon
+              name='Messages'
+              blue
+              styleName={`collab-icon cursor-pointer`}
+              onClick={() => dispatch(messageGroupModerators(group.id)).then(a => a.payload?.data?.messageGroupModerators ? push(`/messages/${a.payload.data.messageGroupModerators}?prompt=${encodeURIComponent(prompt)}`) : null )}
+            />
           </div>
         )
       })}
@@ -64,4 +68,19 @@ const collabText = (group) => {
     support: `${group.name} is seeking farm support`,
     equipment_sharing: `${group.name} is interested in equipment sharing`
   }
+}
+
+// XXX: flag for internationalization/translation
+const promptLookup = {
+  research: 'research projects',
+  events: 'event collaboration',
+  volunteering: 'volunteer opportunties',
+  mentorship: 'mentorship & advice',
+  cooperative: 'cooperatives',
+  buy: 'purchasing from you',
+  markets: 'new markets',
+  ecosystem_service_markets: 'ecosystem services',
+  loans: 'low-cost loans',
+  support: 'farm support',
+  equipment_sharing: 'equipment sharing'
 }
