+++ src/routes/Messages/Messages.js
@@ -19,21 +19,6 @@ import './Messages.scss'
 
 export const NEW_THREAD_ID = 'new'
 
-// XXX: flag for internationalization/translation
-const promptLookup = {
-  research: 'research projects',
-  events: 'event collaboration',
-  volunteering: 'volunteer opportunties',
-  mentorship: 'mentorship & advice',
-  cooperative: 'cooperatives',
-  buy: 'purchasing from you',
-  markets: 'new markets',
-  ecosystem_service_markets: 'ecosystem services',
-  loans: 'low-cost loans',
-  support: 'farm support',
-  equipment_sharing: 'equipment sharing'
-}
-
 export default class Messages extends React.Component {
   static defaultProps = {
     participants: [],
@@ -69,7 +54,7 @@ export default class Messages extends React.Component {
     }
 
     if (prompt) {
-      this.updateMessageText(`Hi there, I'd like to talk about ${promptLookup[prompt]}.`)
+      this.updateMessageText(prompt)
       this.props.changeQuerystringParam(this.props, 'prompt', null)
     }
   }
