+++ src/store/constants.js
@@ -85,6 +85,7 @@ export const MARK_ACTIVITY_READ = 'MARK_ACTIVITY_READ'
 export const MARK_ACTIVITY_READ_PENDING = 'MARK_ACTIVITY_READ_PENDING'
 export const MARK_ALL_ACTIVITIES_READ = 'MARK_ALL_ACTIVITIES_READ'
 export const MARK_ALL_ACTIVITIES_READ_PENDING = 'MARK_ALL_ACTIVITIES_READ_PENDING'
+export const MESSAGE_GROUP_MODERATORS = 'MESSAGE_GROUP_MODERATORS'
 export const PROCESS_STRIPE_TOKEN = 'PROCESS_STRIPE_TOKEN'
 export const PROCESS_STRIPE_TOKEN_PENDING = PROCESS_STRIPE_TOKEN + _PENDING
 export const REJECT_GROUP_RELATIONSHIP_INVITE = 'REJECT_GROUP_RELATIONSHIP_INVITE'
